CMake rewrites (#44)
* Test directory whitespace changes

* Mark move constructors/assign operators noexcept

* Defaulted special member functions are non-throwing

* Move options to front of CMakeLists

* Regularize cmake options

* cmake_dependent_option, not option

* Clean up some cmake

* Update some cmake formatting

* Clean up cmake a bit

* X_ENABLE -> ENABLE_X; AUTOBUILD_ZMQ and _SuiteSparse don't currently do anything

* Propogate ENABLE_KLU better. (also delete griddyn.fmu, why does it keep showing up?)

* KLU_ENABLE -> ENABLE_KLU

* Remove misleading header

* See if Sundials is finding KLU

* Actually see if Sundials is finding KLU this time

* Oops. I checked the rest of the ENABLE_MPIs, they look right

* Does travis get undefined symbol issues with FMI here?

* Attempt to reduce the global state passed through the project

* Add updated FindBoost

* Print slightly more debug information

* Generate libraryLoader when we actually know what to put there

* Fix issues brought up in the PR

* Including all the headers bites again...

* Revert boost cmake changes

* Dump compile commands on travis failure, to assist in debugging

* INTERFACE_SYSTEM_INCLUDE_DIRECTORIES doesn't appear to be transitive

* Add FORCE_DEPENDENCY_REBUILD, turn it on (briefly) for appveyor

* Add updated build options to the installation document

* Tested the updated try_compile commands, tested deleting some of the directory includes. Added an install interface generator for griddyn

* Fix boost includes, fix appveyor cache

* Disable APPVEYOR_CACHE_SKIP_RESTORE, the cache is fixed now

* Resolve comments brought up in the PR

* Use both INTERFACE_INCLUDE_DIRECTORIES (for including at all) and INTERFACE_SYSTEM_INCLUDE_DIRECTORIES (for marking as system) for boost

+++ src/fmi/fmi_models/CymeDistLoad.cpp
@@ -81,7 +81,9 @@ void CymeDistLoadME::loadConfigFile (const std::string &configFileName)
     configFile = configFileName;
     auto mval = doc["models"];
 
-    auto model = mval[configIndex];
+    // this is ambiguous when configIndex is int64_t. unsigned skips the < 0
+    // check, so pass this as signed
+    auto model = mval[static_cast<int>(configIndex)];
     if (model.isObject ())
     {
         auto fmu_path = model["fmu_path"].asString ();
